IMPORT: Graphic Pack
+ ADD: Action and DialogueSystem

+++ Assets/Sources/Interfaces/ITakeFix.cs
@@ -1,4 +1,4 @@
 public interface ITakeFix 
 {
-    void TakeFix(float fixForce);
+    void TakeFix(int fixForce);
 }
